+++ CHANGELOG.md
@@ -7,8 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Changed
-- CPX-AP: datatype of argument channel in read_isda() and write_isda() from int to list[int] to pass a list of channels which should be r/w
-- CPX-AP: return value of read_isda() from any to list[any]
+- CPX-AP: datatype of argument channel in read_isda() and write_isda() from int to (list[int] and int) to pass a list of channels or a single channel which should be r/w
+- CPX-AP: return value of read_isda() from any to list[any] and any
 
 ## v0.9.0 - 05.05.25
 ### Added